// frontend/src/api/generated/models/projectThermal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */
import type { OrientationCompliance } from "./orientationCompliance";
import type { ProjectThermalThermalWallAreas } from "./projectThermalThermalWallAreas";
import type { ProjectThermalThermalZone } from "./projectThermalThermalZone";
import type { ThermalPosition } from "./thermalPosition";
import type { ThermalUseEnum } from "./thermalUseEnum";
import type { ThermalVerdictEnum } from "./thermalVerdictEnum";

export interface ProjectThermal {
  project_id: string;
  thermal_zone: (typeof ProjectThermalThermalZone)[keyof typeof ProjectThermalThermalZone] | null;
  thermal_use: ThermalUseEnum;
  /** @nullable */
  thermal_wall_areas: ProjectThermalThermalWallAreas;
  positions: ThermalPosition[];
  orientations: OrientationCompliance[];
  verdict: ThermalVerdictEnum;
}